server/requirements.txt
asgiref==3.5.2
aspose-words==22.11.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
dropbox==11.36.0
idna==3.4
ply==3.11
pycparser==2.21
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stone==3.3.1
tzdata==2022.6
urllib3==1.26.12
